python/requirements.txt
fastapi
uvicorn
web3
python-dotenv
pyaudio
SpeechRecognition


# Gemini / Google Generative AI SDK
google-generativeai>=0.4.1

# Voice recording and transcription
sounddevice>=0.4.6
scipy>=1.11.0
SpeechRecognition>=3.10.0

# Optional but recommended
python-dotenv>=1.0.1

# For numpy used in voice normalization
numpy>=1.24.0
